tests/wpt/web-platform-tests/css/css-box/parsing/min-height-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing min-height with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-min-height">
<meta name="assert" content="min-height supports only the grammar '[ [<length> | <percentage>] && [border-box | content-box]? ] | available | min-content | max-content | fit-content'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("min-height", "complex");
test_invalid_value("min-height", "none");
test_invalid_value("min-height", "none available");
test_invalid_value("min-height", "max-content 10px");
test_invalid_value("min-height", "20% available");

test_invalid_value("min-height", "-10px");
test_invalid_value("min-height", "-20%");
test_invalid_value("min-height", "60");
test_invalid_value("min-height", "10px 20%");
test_invalid_value("min-height", "content-box border-box");
test_invalid_value("min-height", "10px border-box 20%");
test_invalid_value("min-height", "content-box 20% border-box");

// The following is not yet rejected by browsers:
test_invalid_value("min-height", "auto");
</script>
</body>
</html>
